+++ lib/components/text_components/kills.dart
@@ -19,7 +19,7 @@ class KillInfo extends TextComponent with HasGameRef<EilemitweileGame> {
   Future<void>? onLoad() {
     text_content = "Kills\nRed: 0\nBlue: 0\nGreen: 0\nPurple: 0";
     anchor = Anchor.center;
-    position.setValues(175, 1350);
+    position.setValues(260, 1300);
     text = text_content;
 
     return super.onLoad();