+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## v0.0.7
+- Fix: Restart game when someone has won
+- Violet to purple
+- Block bench to heaven
+- Fix: 12 to heaven
+- Fix: Center dice numbers
+
+## v0.0.6
+- Restart game when three players are finished
+- Fix: AI didn't really look kor killing opportunities
+
 ## v0.0.5
 - Fix: arrange tokens on ladders
 - AI: a bit more intelligent, checks if a token can move on a bench or kill